Improve collaborators and owner columns in shared file lists
When sharing with multiple targets, they must all be listed there
instead of a single entry. This is fixed by replacing the code that make
share entries unique into an aggregation that has been inspired by the
one used in OC core.
Also includes group shares and link shares now.

Added avatars and icons for the displayed collaborators in the column.