Fix video loading in media viewer on public pages
The `mediaSource` plugin doesn't support signed urls at the moment,
because it would slow down the whole file list init the way signed urls
are implement at the moment. Reason: signed urls retrieve the signing
key on every request. As a result ALL requests for e.g. thumbnails in
the files list will make two requests instead of one. As long as that is
not fixed, we will continue to fetch previews and images in general
(also in the mediaviewer) as blobs. Only other file types (e.g. videos)
are fetched with signed urls.