Prevent binding only to loopback IP when running in watch mode.
This is required when running the acceptance tests on Windows, it allows the selenium docker containers to access the frontend due to the host binding in rollup (when running `yarn serve`).  I have considered the implications but can't see any downside to having this as the default, it doesn't break any existing functionality.