feat(eventbus): resource loading happens multiple times because of multiple event subscriptions to the same topic
the eventbus is able to subscribe many times to the same topic, this
could be problematic in cases where the subscription is build on
component mount hook.

This leads to multiple unwanted subscriptions on
the same topic, which results in unnecessary requests and state
changes.

This is fixed by allowing the eventBus subscriptions to
unsubscribe on a topic and token.