Introduce individual sorting query params for share tables
The shared with me page has two different tables: one for the pending
shares (if any) and one for the accepted or declined shares (depending
on the view mode). This commit introduces query params for sorting
(sortBy and sortDir) for each table. Before they were both sorted by the
same query params.