Introduce persisted route query composable
It's being used to auto-wire RouteQuery params to local storage. It also
enforced RouteQuery params to be set in the URL, even if the route was
visited without them. With this we start to always have a full state in
the URL.
Started with the usePagination composable which now handles the items per page
internally, including persistence to localStorage. Later on this can be
replaced with another storage (like settings in the backend).