Introduce Shares overview page
Introduce `active` property for navItems in shares navigation

The `active` property that's being used in the template was simply
missing in the navItems. Added it and used the chance to switch over to
composition API for the new SharesNavigation component.

Use $gettext in template directly

Not needed to define it in setup function

Finalize SharesNavigation component && adjust tests