[full-ci] More app top bar fixes (#7362)
* Move Resource interface to web-client

* Add getFileResource helper to AppFileHandling composable

* web-app-external: use getFileResource helper

* Fix permissions type in Resource interface

* web-app-pdf-viewer: make AppTopBar always visible

* web-app-pdf-viewer: use getFileResource helper

* web-app-text-editor: make AppTopBar always visible

* web-app-text-editor: add error display

* web-app-text-editor: use getFileResource helper

* web-app-preview: simplify raw file loading

... by moving complexity to useAppFileHandling composable

* web-app-pdf-viewer: Use getUrlForResource

* web-app-preview: Use revokeUrl from file handling composable

* Add current PR to unreleased changelog item

* useAppFileHandling.getUrlForResource: add disposition option

which can be either inline or attachment and can be used to define the purpose of the retrieved url.
Chrome allows us to use signed/downloadURLs for video playback although they send a Content-Disposition header, so we can just use a signed url and don't need to create a blob url for this.
We cannot embed PDFs in the same way, as they trigger downloads. So we need to create a blob url for those.

If you need a download url, it's fine (or even desired) for the response
to contain the Content-Disposition header.