refactor: quota modal usage
Before, the state of the quota modal was handled by the outside where it was being used. This caused quite some boilerplate code because it's effectively always the same.

Refactors the quota modal usage so that the modal state handling is done via the quota action instead, meaning the consuming party doesn't need to care about it anymore.