chore: update cross-fetch to v3.1.8 and node-fetch to v2.7.0
Dependency node-fetch has been updated to v2.7.0. We found out that
cucumber-js is currently incompatible with node-fetch v3, so we can't
upgrade to node-fetch v3 at the moment.
Dependency cross-fetch is a peer dependency of the owncloud-sdk and has
been added to our root level package.json to get an up to date v3 of
cross-fetch to mitigate security issues.