refactor: remove highlightedFile getter from store
Removes the `highlightedFile` getter from the vuex file store. The reason behind this decision is that the getter was really opaque to use as it sometimes referred to the current folder and sometimes to the first selected file (which is already confusing in itself since the user can highlight more than 1 resources).