fix: idp logout issues
Fixes falsely showing the logout page after opening ownCloud Web with an expired token. The issue seemed to be related to the fact that we manually logged out the user after token expiry, whereas the correct way is to let the IDP auth flow handle this. There is no need for a custom implementation on our side.