feat: move access token timer to web worker
Introduces a web worker for the access token timer to run in. As soon as a token is about to expire, the web worker will inform the main thread that it needs to retrieve a new token from the server.

This ensures that the timer runs es expected even if the browser tab is being throttled or suspended. As a result, long running operation such as uploads or move actions can continue in the background.