fix: prevent apps from re-initializing on logo click
Prevents all apps from re-initializing when clicking the ownCloud logo on public link pages.

The issue happens because clicking the logo navigates the user to the `resolvePublicLink` page, which results in the public context being reset. We can omit this reset in that case though because we can guarantee that the user will either land on a public page, or on some other route which then resets the context anyways.