feat: add file id support to webdav client
Adds support for file ids to the webdav client. This currently excludes delete and restore operations because the server lacks their support at the moment.

Id-based webdav operations perform much better in general since the server doesn't need to traverse file paths.

Note that all webdav operations using a file id won't return any path information. Meaning if you need to work with a resource path, it's recommended to keep providing a path to the client.

The implementation prioritizes a path over a file id if both parameters are given.