refactor: transform graph roles to object
Refactors the unified graph role definitions from an array to an object because it simplifies the way how to work with those. When retrieving a role via its id (which is 99% if the use cases), it can now be done with the object index instead of looping through the array.