chore: remove font-awesome dependency
As far as I can see, the dependency was only used to display the attribution in the browser console. However, this attribution has been removed with their version `6.6.0`.

Also, according to https://fontawesome.com/license/free it's sufficient to have the attribution embedded in the downloaded icons (which we have).