fix: tailwind reference (#1086)
`@reference 'tailwindcss'` will only reference the default tailwind styles, not the custom ones we have in `tailwind.css`. Instead, the css file should be referenced. However, since this is really cumbersome to do in every component, this adds a custom vite plugin that replaces `@reference 'tailwindcss'` with the correct path.